requirements.txt
gradio
torch
wikipedia
tavily-python
sentence-transformers
transformers
diskcache
